feat: add handling of removed token on backoffice
